pyproject.toml
[project]
name = "belegmanager"
version = "0.1.0"
description = "Lokale Belegverwaltung fuer selbststaendige Kuenstler:innen"
readme = "README.md"
license = {text = "AGPL-3.0-or-later"}
classifiers = [
  "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12"
]
requires-python = ">=3.12"
dependencies = [
  "nicegui>=2.7.0",
  "sqlmodel>=0.0.22",
  "pillow>=10.4.0",
  "python-dateutil>=2.9.0",
  "pypdfium2>=4.30.0",
  "pillow-heif>=0.21.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["belegmanager*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
